pyproject.toml
[tool.poetry]
name = "7b-sqlmaster"
version = "0.1.0"
description = ""
authors = ["jjovalle <jj.ovalle@uniandes.edu.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
streamlit = "^1.31.1"
pandas = "^2.2.1"
langchain = "^0.1.11"
python-dotenv = "^1.0.1"
pandasql = "^0.7.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"
langchain-openai = "^0.0.8"
isort = "^5.13.2"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
